+++ PlayerEndpoint.go
@@ -43,12 +43,20 @@ func (elem *PlayerEndpoint) Play() error {
 	}
 
 	// Call server and wait response
-	response := <-elem.connection.Request(req)
-
-	// Returns error or nil
-	if response.Error != nil {
-		return errors.New(fmt.Sprintf("[%d] %s %s", response.Error.Code, response.Error.Message, response.Error.Data))
+	responses, err := elem.connection.Request(req)
+	if err != nil {
+		return err
 	}
+	select {
+	case response := <-responses:
+		// Returns error or nil
+		if response.Error != nil {
+			return errors.New(fmt.Sprintf("[%d] %s %s", response.Error.Code, response.Error.Message, response.Error.Data))
+		}
+	case <-elem.connection.closeSig:
+		return ErrConnectionClosing
+	}
+
 	return nil
 
 }