+++ WebrtcEndpoint.go
@@ -49,13 +49,21 @@ func (elem *WebRtcEndpoint) GatherCandidates() error {
 	}
 
 	// Call server and wait response
-	response := <-elem.connection.Request(req)
-
-	// Otherwise we want to wait for the other candidates
-	// Returns error or nil
-	if response.Error != nil {
-		return errors.New(fmt.Sprintf("[%d] %s %s", response.Error.Code, response.Error.Message, response.Error.Data))
+	responses, err := elem.connection.Request(req)
+	if err != nil {
+		return err
 	}
+	select {
+	case response := <-responses:
+		// Otherwise we want to wait for the other candidates
+		// Returns error or nil
+		if response.Error != nil {
+			return errors.New(fmt.Sprintf("[%d] %s %s", response.Error.Code, response.Error.Message, response.Error.Data))
+		}
+	case <-elem.connection.closeSig:
+		return ErrConnectionClosing
+	}
+
 	return nil
 }
 
@@ -77,12 +85,20 @@ func (elem *WebRtcEndpoint) AddIceCandidate(candidate IceCandidate) error {
 	}
 
 	// Call server and wait response
-	response := <-elem.connection.Request(req)
-
-	// Returns error or nil
-	if response.Error != nil {
-		return errors.New(fmt.Sprintf("[%d] %s %s", response.Error.Code, response.Error.Message, response.Error.Data))
+	responses, err := elem.connection.Request(req)
+	if err != nil {
+		return err
 	}
+	select {
+	case response := <-responses:
+		// Returns error or nil
+		if response.Error != nil {
+			return errors.New(fmt.Sprintf("[%d] %s %s", response.Error.Code, response.Error.Message, response.Error.Data))
+		}
+	case <-elem.connection.closeSig:
+		return ErrConnectionClosing
+	}
+
 	return nil
 
 }